separate system roots test from security connector
- each OS will have a separate test for its system roots, since they
each require different implementations
- formatting script hasn't run yet
