+++ build.yaml
@@ -2758,6 +2758,16 @@ targets:
   - gpr_test_util
   - gpr
   uses_polling: false
+- name: grpc_linux_system_roots_test
+  build: test
+  language: c
+  src:
+  - test/core/security/system_roots_test_linux.cc
+  deps:
+  - grpc_test_util
+  - grpc
+  - gpr_test_util
+  - gpr
 - name: grpc_print_google_default_creds_token
   build: tool
   language: c
